+++ assets/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 19-8-21
 - Moved profile picture and changelog to `assets/`
 - Added comment to `fumo_mode.py` explaning what happened.
+- Added genres to `$anisearch`
 - Changed NSFW error message to an embed
   - Added Sagiri picture just because I thought it would be funny